moving the class into proper place
