+++ src/Modules/ApiGeneratorModule/ApiController/ApiHooks.php
@@ -1,6 +1,5 @@
 <?php
-
-namespace Crocodicstudio\Crudbooster\Controllers\ApiController;
+namespace Crocodicstudio\Crudbooster\Modules\ApiGeneratorModule\ApiController;
 
 trait ApiHooks
 {
